+++ packages/video_player/video_player_avfoundation/lib/src/video_player_plugin_messages.g.dart
@@ -134,15 +134,72 @@ class PlatformVideoViewCreationParams {
   int get hashCode => _deepHash(<Object?>[runtimeType, ..._toList()]);
 }
 
+/// Pigeon equivalent of video_player_avfoundation's FairPlayDrmConfiguration.
+class PlatformFairPlayDrmConfiguration {
+  PlatformFairPlayDrmConfiguration({
+    required this.certificateUri,
+    required this.licenseUri,
+    required this.licenseHeaders,
+    this.contentId,
+  });
+
+  String certificateUri;
+
+  String licenseUri;
+
+  Map<String, String> licenseHeaders;
+
+  String? contentId;
+
+  List<Object?> _toList() {
+    return <Object?>[certificateUri, licenseUri, licenseHeaders, contentId];
+  }
+
+  Object encode() {
+    return _toList();
+  }
+
+  static PlatformFairPlayDrmConfiguration decode(Object result) {
+    result as List<Object?>;
+    return PlatformFairPlayDrmConfiguration(
+      certificateUri: result[0]! as String,
+      licenseUri: result[1]! as String,
+      licenseHeaders: (result[2]! as Map<Object?, Object?>).cast<String, String>(),
+      contentId: result[3] as String?,
+    );
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  bool operator ==(Object other) {
+    if (other is! PlatformFairPlayDrmConfiguration || other.runtimeType != runtimeType) {
+      return false;
+    }
+    if (identical(this, other)) {
+      return true;
+    }
+    return _deepEquals(certificateUri, other.certificateUri) &&
+        _deepEquals(licenseUri, other.licenseUri) &&
+        _deepEquals(licenseHeaders, other.licenseHeaders) &&
+        _deepEquals(contentId, other.contentId);
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  int get hashCode => _deepHash(<Object?>[runtimeType, ..._toList()]);
+}
+
 class CreationOptions {
-  CreationOptions({required this.uri, required this.httpHeaders});
+  CreationOptions({required this.uri, required this.httpHeaders, this.fairPlayDrm});
 
   String uri;
 
   Map<String, String> httpHeaders;
 
+  PlatformFairPlayDrmConfiguration? fairPlayDrm;
+
   List<Object?> _toList() {
-    return <Object?>[uri, httpHeaders];
+    return <Object?>[uri, httpHeaders, fairPlayDrm];
   }
 
   Object encode() {
@@ -154,6 +211,7 @@ class CreationOptions {
     return CreationOptions(
       uri: result[0]! as String,
       httpHeaders: (result[1]! as Map<Object?, Object?>).cast<String, String>(),
+      fairPlayDrm: result[2] as PlatformFairPlayDrmConfiguration?,
     );
   }
 
@@ -166,7 +224,9 @@ class CreationOptions {
     if (identical(this, other)) {
       return true;
     }
-    return _deepEquals(uri, other.uri) && _deepEquals(httpHeaders, other.httpHeaders);
+    return _deepEquals(uri, other.uri) &&
+        _deepEquals(httpHeaders, other.httpHeaders) &&
+        _deepEquals(fairPlayDrm, other.fairPlayDrm);
   }
 
   @override
@@ -221,11 +281,14 @@ class _PigeonCodec extends StandardMessageCodec {
     } else if (value is PlatformVideoViewCreationParams) {
       buffer.putUint8(129);
       writeValue(buffer, value.encode());
-    } else if (value is CreationOptions) {
+    } else if (value is PlatformFairPlayDrmConfiguration) {
       buffer.putUint8(130);
       writeValue(buffer, value.encode());
-    } else if (value is TexturePlayerIds) {
+    } else if (value is CreationOptions) {
       buffer.putUint8(131);
+      writeValue(buffer, value.encode());
+    } else if (value is TexturePlayerIds) {
+      buffer.putUint8(132);
       writeValue(buffer, value.encode());
     } else {
       super.writeValue(buffer, value);
@@ -238,8 +301,10 @@ class _PigeonCodec extends StandardMessageCodec {
       case 129:
         return PlatformVideoViewCreationParams.decode(readValue(buffer)!);
       case 130:
-        return CreationOptions.decode(readValue(buffer)!);
+        return PlatformFairPlayDrmConfiguration.decode(readValue(buffer)!);
       case 131:
+        return CreationOptions.decode(readValue(buffer)!);
+      case 132:
         return TexturePlayerIds.decode(readValue(buffer)!);
       default:
         return super.readValueOfType(type, buffer);