+++ packages/video_player/video_player_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.13.0
+
+* Adds `WidevineDrmConfiguration`, allowing playback of Widevine-protected
+  network streams via Media3's DRM support.
+
 ## 2.12.0
 
 * Fixes a [bug](https://github.com/flutter/flutter/issues/176575) where some videos report an incorrect duration when initialized without a video duration.